perf(data): batch yfinance historical downloads to cut 401 Invalid Crumb
- get_historical_data_batch: agrupa cache-misses en yf.download por chunk
  (un crumb compartido); refactor a helpers _read/_write_historical_cache,
  _normalize_ohlcv, _finalize_historical (single sin cambios de comportamiento)
- cablea 6 callers multi-ticker: portfolio_backtest, harness_walkforward,
  run_switcher_validation, run_cross_sectional_validation, rsi_scanner,
  portfolio_tab; prefetch_harness_cache con flag -b
- tests/test_historical_batch.py (7 casos)

Suite: 852 passed, 1 skipped

+++ tests/test_historical_batch.py
@@ -134,8 +134,9 @@ def test_multiindex_split_per_ticker(monkeypatch):
 def test_partial_failure_isolated(monkeypatch, _isolate):
     monkeypatch.setattr(yf_mod, "_read_historical_cache", lambda *a: None)
     good = _ohlcv()
-    nan_frame = _ohlcv()
-    nan_frame[:] = float("nan")  # símbolo deslistado → Yahoo rellena con NaN
+    # símbolo deslistado → Yahoo rellena con NaN (float para evitar el cast int64)
+    nan_frame = _ohlcv().astype("float64")
+    nan_frame[:] = float("nan")
     monkeypatch.setattr(
         yf_mod, "_download_batch",
         lambda chunk, p, i: _multiindex_batch({"AAPL": good, "ZZZZ": nan_frame}),
